requirements.txt
Django==4.1.7
simplejson==3.18.4
djangorestframework==3.14.0
requests==2.28.2
urllib3==1.26.15
djangorestframework-simplejwt==5.2.2